+++ src/store/index.ts
@@ -1,2 +1,4 @@
 export * from "./database";
+export * from "./errors";
 export * from "./schema";
+export * from "./syx-codec";